requirements.txt
python-fasthtml>=0.1.0
pillow>=9.0.0
pytesseract>=0.3.8
numpy>=1.22.0
pandas>=1.4.0
scikit-learn>=1.0.0
controlflow>=0.1.0
langchain-mistralai>=0.0.3
